+++ src/parsers/qqmusic.rs
@@ -51,15 +51,15 @@ impl IParsers for QQMusicParser {
                 (None, None)         => break,
             };
             let d1 = content[cpos..d1_end]
-                .parse::<u32>()
+                .parse::<u16>()
                 .map_err(|e| format!("d1: {:?} raw={:?}", e, &content[cpos..d1_end]))?;
 
             let Some(rp) = memchr(b')', &cbytes[cpos..]) else { break };
             cpos += rp + 1;
 
             result.push(TextInfo {
-                start_time: self.get_offset_time(s, s1)? as u16,
-                duration: d1 as u16,
+                start_time: self.get_offset_time(s, s1)?,
+                duration: d1,
                 text: text_raw,
             });
         }
